Read animation movement v0/v1 as floats

+++ ValveResourceFormat/Resource/ResourceTypes/ModelAnimation/AnimationMovement.cs
@@ -76,8 +76,8 @@ namespace ValveResourceFormat.ResourceTypes.ModelAnimation
         {
             EndFrame = frameBlock.GetInt32Property("endframe");
             MotionFlags = (ModelAnimationMotionFlags)frameBlock.GetInt32Property("motionflags");
-            V0 = frameBlock.GetInt32Property("v0");
-            V1 = frameBlock.GetInt32Property("v1");
+            V0 = frameBlock.GetFloatProperty("v0");
+            V1 = frameBlock.GetFloatProperty("v1");
             Angle = frameBlock.GetFloatProperty("angle");
             Vector = new Vector3(frameBlock.GetFloatArray("vector"));
             Position = new Vector3(frameBlock.GetFloatArray("position"));
